Move GoogleAPI endpoint to its own file
