+++ app/api.py
@@ -6,6 +6,7 @@ from fastapi import APIRouter
 
 from app.core import endpoints_core
 from app.core.auth import endpoints_auth
+from app.core.google_api import endpoints_google_api
 from app.core.groups import endpoints_groups
 from app.core.notification import endpoints_notification
 from app.core.payment import endpoints_payment
@@ -18,6 +19,7 @@ api_router = APIRouter()
 
 api_router.include_router(endpoints_auth.router)
 api_router.include_router(endpoints_core.router)
+api_router.include_router(endpoints_google_api.router)
 api_router.include_router(endpoints_groups.router)
 api_router.include_router(endpoints_notification.router)
 api_router.include_router(endpoints_payment.router)
